+++ apps/block_scout_web/test/block_scout_web/controllers/api/v2/smart_contract_controller_test.exs
@@ -461,8 +461,7 @@ defmodule BlockScoutWeb.API.V2.SmartContractControllerTest do
           "outputs" => [],
           "name" => "disableWhitelist",
           "inputs" => [%{"type" => "bool", "name" => "disable", "internalType" => "bool"}]
-        },
-        %{"type" => "fallback"}
+        }
       ]
 
       target_contract = insert(:smart_contract, abi: abi)
@@ -496,8 +495,6 @@ defmodule BlockScoutWeb.API.V2.SmartContractControllerTest do
                "inputs" => [%{"type" => "address", "name" => "_address", "internalType" => "address"}],
                "method_id" => "c683630d"
              } in response
-
-      assert %{"type" => "fallback"} in response
     end
 
     test "get array of addresses within read-methods", %{conn: conn} do